backend/requirements.txt
fastapi
uvicorn
sqlalchemy
scikit-learn
textblob
matplotlib
wordcloud
selenium
groq
python-dotenv
python-multipart
gunicorn
sqlalchemy
pydantic
python-dotenv
selenium
crawl4ai
openai
passlib
bcrypt
python-multipart
matplotlib
plotly
beautifulsoup4
html2text
groq
